move out volume 2 volume functions
