+++ rts/Sim/CMakeLists.txt
@@ -13,6 +13,7 @@ add_library(engineSim STATIC
 		"${CMAKE_CURRENT_SOURCE_DIR}/Misc/BuildingMaskMap.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Misc/CategoryHandler.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Misc/CollisionHandler.cpp"
+		"${CMAKE_CURRENT_SOURCE_DIR}/Misc/CollisionHandlerVV.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Misc/CollisionVolume.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Misc/CommonDefHandler.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Misc/DamageArray.cpp"
