+++ test/CMakeLists.txt
@@ -558,7 +558,8 @@ endif()
 			# "${CMAKE_CURRENT_SOURCE_DIR}/engine/Sim/Misc/CollisionHandlerBoxVolumeSupport.cpp"
 			# "${CMAKE_CURRENT_SOURCE_DIR}/engine/Sim/Misc/CollisionVolumeTestSupport.cpp"
 			"${CMAKE_CURRENT_SOURCE_DIR}/engine/Sim/Misc/CollisionHandlerSupport.cpp"
-			"${ENGINE_SOURCE_DIR}/Sim/Misc/CollisionHandler.cpp"
+			# "${ENGINE_SOURCE_DIR}/Sim/Misc/CollisionHandler.cpp"
+			"${ENGINE_SOURCE_DIR}/Sim/Misc/CollisionHandlerVV.cpp"
 			"${ENGINE_SOURCE_DIR}/Sim/Misc/CollisionVolume.cpp"
 			"${ENGINE_SOURCE_DIR}/System/Matrix44f.cpp"
 			"${ENGINE_SOURCE_DIR}/System/Quaternion.cpp"
